refactor layout, image lazy loadinginit
