+++ components/core/years/small-grid.tsx
@@ -13,11 +13,19 @@ const SmallGrid = (props: IGridProps) => {
           minHeight={["176px", "176px", "360px"]}
           height="100%"
           objectFit={"cover"}
+          loading="lazy"
         />
       </Box>
       <VStack width="176px" height="100%">
         {changelogs.slice(1, changelogs.length).map(({ imageUrl, slug }, index) => (
-          <Image key={index} src={imageUrl} alt={slug} height="100%" objectFit={"cover"} />
+          <Image
+            key={index}
+            src={imageUrl}
+            alt={slug}
+            height="100%"
+            objectFit={"cover"}
+            loading="lazy"
+          />
         ))}
       </VStack>
     </HStack>
