+++ components/layout/index.tsx
@@ -1,66 +1,30 @@
-import React, { ReactNode } from "react";
-import Link from "next/link";
+import React from "react";
 import Head from "next/head";
 import { defaultPx } from "lib/utils/default-container-px";
 import TryBanner from "components/core/try-banner";
 import Navbar from "components/core/navbar";
 import { Footer } from "components/core/footer";
-import { Box, Button, Container, Heading, HStack, Text, VStack } from "@chakra-ui/react";
+import { Box, Container, Heading, HStack, Text, VStack } from "@chakra-ui/react";
 import TimeSelectionTabs from "../core/time-selection-tabs";
-import useTimelineStore from "lib/state/use-timeline-store";
 import { motion } from "framer-motion";
 
-export interface MainLayoutProps {
-  page?: number;
-  children: ReactNode;
-  itemsPerPage?: number;
-  totalItems?: {
-    weeks: number;
-    months: number;
-    years: number;
-  };
-  infiniteScrollingView?: "year" | "month";
+export interface ILayoutProps {
+  children: React.ReactNode;
 }
 
-export const MainLayout = ({
-  page,
-  children,
-  itemsPerPage,
-  totalItems,
-  infiniteScrollingView,
-}: MainLayoutProps) => {
-  const metaTitle = `${
-    infiniteScrollingView ? "" : page > 0 ? `Page ${page} -` : ""
-  } June Changelog`;
-  const timeline = useTimelineStore();
-
-  React.useEffect(() => {
-    const hash = window?.location.hash ?? "";
-
-    timeline.setView(
-      hash ? (hash === "#months" ? "months" : hash === "#years" ? "years" : "weeks") : "weeks"
-    );
-  }, []);
-
-  const hasMorePage =
-    !infiniteScrollingView && page < Math.floor(totalItems[timeline.view] / itemsPerPage);
-
+export default function Layout({ children }: ILayoutProps) {
   return (
     <>
       <Head>
-        <title>{metaTitle}</title>
         <link rel="icon" href="/favicon.ico" />
-        <meta name="title" content={metaTitle} />
         <meta name="description" content="Discover new updates and improvements to June." />
         <meta name="image" content="https://changelog.june.so/social.png" />
         <meta property="og:type" content="website" />
         <meta property="og:url" content="https://changelog.june.so" />
-        <meta property="og:title" content={metaTitle} />
         <meta property="og:description" content="Discover new updates and improvements to June." />
         <meta property="og:image" content="https://changelog.june.so/social.png" />
         <meta name="twitter:card" content="summary_large_image" />
         <meta name="twitter:url" content="https://changelog.june.so" />
-        <meta name="twitter:title" content={metaTitle} />
         <meta name="twitter:description" content="Discover new updates and improvements to June." />
         <meta name="twitter:image" content="https://changelog.june.so/social.png" />
         <link
@@ -80,18 +44,20 @@ export const MainLayout = ({
           <Navbar />
         </motion.div>
         <motion.div
-          variants={{
-            hidden: { opacity: 0 },
-            visible: { opacity: 1, transition: { duration: 0.6, delay: 0.2 } },
-          }}
-          layout
-          transition={{ duration: 0 }}
+          // variants={{
+          //   hidden: { opacity: 0 },
+          //   visible: { opacity: 1, transition: { duration: 0.6, delay: 0.2 } },
+          // }}
+          // layout
+          // transition={{ duration: 0 }}
           style={{
             position: "sticky",
             top: "32px",
             zIndex: 1,
             paddingBottom: "32px",
+            opacity: 1,
           }}
+          // layout
         >
           <TimeSelectionTabs />
         </motion.div>
@@ -118,40 +84,6 @@ export const MainLayout = ({
                   </VStack>
                 </VStack>
               </motion.div>
-              <motion.div
-                hidden={!!infiniteScrollingView}
-                variants={{
-                  hidden: { opacity: 0, y: 20 },
-                  visible: { opacity: 1, y: 0, transition: { duration: 0.6, delay: 0.3 } },
-                }}
-              >
-                <VStack align={["stretch", "stretch", "center"]}>
-                  {page === 0 && hasMorePage ? (
-                    <Link href={`/page/1#${timeline.view}`}>
-                      <Button variant="landingOutline" size="landingLg">
-                        Load more
-                      </Button>
-                    </Link>
-                  ) : (
-                    <HStack justifyContent="center" spacing={4}>
-                      {page > 0 && (
-                        <Link href={`/page/${page - 1}${"#" + timeline.view}`}>
-                          <Button variant="landingOutline" size="landingLg">
-                            Previous page
-                          </Button>
-                        </Link>
-                      )}
-                      {hasMorePage && (
-                        <Link href={`/page/${page + 1}${"#" + timeline.view}`}>
-                          <Button variant="landingOutline" size="landingLg">
-                            Next page
-                          </Button>
-                        </Link>
-                      )}
-                    </HStack>
-                  )}
-                </VStack>
-              </motion.div>
             </VStack>
           </Container>
           <motion.div
@@ -168,4 +100,4 @@ export const MainLayout = ({
       </motion.div>
     </>
   );
-};
+}
