+++ public/rss.xml
@@ -4,7 +4,7 @@
         <title>June Changelog</title>
         <link>https://changelog.june.so</link>
         <description>How June gets better every week</description>
-        <lastBuildDate>Thu, 20 Jul 2023 10:38:30 GMT</lastBuildDate>
+        <lastBuildDate>Fri, 21 Jul 2023 08:20:40 GMT</lastBuildDate>
         <docs>https://validator.w3.org/feed/docs/rss2.html</docs>
         <generator>Feed for June changelog</generator>
         <image>
